requirements.txt
matrix-nio[e2e]>=0.21.0
aiosqlite>=0.20.0
apscheduler>=3.10.0
pyyaml>=6.0
fastapi>=0.111.0
uvicorn>=0.30.0
httpx>=0.27.0
